extract virtual agent token hook
